Complete tutorial
In this commit you will find the fireball charge change so you can now fire with the left mouse button. The HUD has been updated to include transparancy. An old version of the HUD was still present so I deleted it. Finally, the second part of the tutorial has been completed.